Added CRUD operations for the Artist Portfolio and Skills
